Move memory_eq_* functions to memory_eq_encoder.rs
