+++ prusti-viper/src/encoder/mod.rs
@@ -17,6 +17,7 @@ mod loop_encoder;
 mod mir_encoder;
 mod mir_successor;
 mod mir_interpreter;
+mod memory_eq_encoder;
 mod optimizer;
 mod places;
 mod procedure_encoder;
